mode normlization of devices sever side
