+++ src/web/api/devices.ts
@@ -2,20 +2,21 @@ import { Express } from 'express';
 import { constants } from 'http2';
 import { Device } from 'zigbee-herdsman/dist/controller/model';
 import { ZigbeeNTHomebridgePlatform } from '../../platform';
+import { normalizeDeviceModel } from '../common/utils';
 
 export function mapDevicesRoutes(express: Express, platform: ZigbeeNTHomebridgePlatform) {
   express.get('/api/devices', (_req, res) => {
     const devices: Device[] = platform.zigBeeClient.getAllPairedDevices();
     res.status(constants.HTTP_STATUS_OK);
     res.contentType('application/json');
-    res.end(JSON.stringify({ devices }));
+    res.end(JSON.stringify({ devices: devices.map(device => normalizeDeviceModel(device)) }));
   });
 
   express.get('/api/devices/:ieeeAddr', (req, res) => {
     const device: Device = platform.zigBeeClient.getDevice(req.params.ieeeAddr);
     res.status(constants.HTTP_STATUS_OK);
     res.contentType('application/json');
-    res.end(JSON.stringify({ device }));
+    res.end(JSON.stringify({ device: normalizeDeviceModel(device) }));
   });
 
   express.delete('/api/devices/:ieeeAddr', async (req, res) => {
@@ -24,7 +25,7 @@ export function mapDevicesRoutes(express: Express, platform: ZigbeeNTHomebridgeP
       await platform.unpairDevice(req.params.ieeeAddr);
       res.status(constants.HTTP_STATUS_OK);
       res.contentType('application/json');
-      res.end(JSON.stringify({ device }));
+      res.end(JSON.stringify({ device: normalizeDeviceModel(device) }));
     } else {
       res.status(constants.HTTP_STATUS_NOT_FOUND);
       res.end();
