+++ src/web/ui/actions/devices.ts
@@ -1,41 +1,11 @@
 import { handleError } from './utils';
 import { DeviceState } from '../../../zigbee/types';
+import { DeviceModel } from '../../common/types';
 
 export interface BaseResponse {
   result: 'success' | 'error';
   error?: string;
 }
-
-export type Endpoint = {
-  ID: number;
-  profileID: number;
-  deviceID: number;
-  inputClusters: number[];
-  outputClusters: number[];
-  deviceNetworkAddress: number;
-  deviceIeeeAddress: string;
-  clusters: {
-    [k: string]: {
-      attributes: any;
-    };
-  };
-  _binds: any[];
-  meta: any;
-};
-
-export type DeviceModel = {
-  type: string;
-  ieeeAddr: string;
-  networkAddress: number;
-  manufacturerID: string;
-  manufacturerName: string;
-  powerSource: string;
-  modelID: string;
-  interviewCompleted: boolean;
-  lastSeen: number;
-  softwareBuildID: string;
-  endpoints?: Endpoint[];
-};
 
 export interface DeviceResponse extends BaseResponse {
   devices?: DeviceModel[];
@@ -47,22 +17,6 @@ export interface StateResponse extends BaseResponse {
   state?: DeviceState;
 }
 
-export function normalizeDeviceModel(d): DeviceModel {
-  return {
-    type: d._type,
-    ieeeAddr: d._ieeeAddr,
-    networkAddress: d._networkAddress,
-    manufacturerID: d._manufacturerID,
-    manufacturerName: d._manufacturerName,
-    powerSource: d._powerSource,
-    modelID: d._modelID,
-    interviewCompleted: d._interviewCompleted,
-    softwareBuildID: d._softwareBuildID,
-    lastSeen: d._lastSeen,
-    endpoints: d._endpoints,
-  };
-}
-
 export class DevicesService {
   static async fetchDevices(): Promise<DeviceResponse> {
     try {
@@ -71,7 +25,7 @@ export class DevicesService {
         const json = await response.json();
         return {
           result: 'success',
-          devices: json.devices.map(normalizeDeviceModel),
+          devices: json.devices,
           total: json.devices.length,
         };
       } else {
@@ -89,7 +43,7 @@ export class DevicesService {
         const json = await response.json();
         return {
           result: 'success',
-          device: normalizeDeviceModel(json.device),
+          device: json.device,
         };
       } else {
         return handleError(await response.text());
@@ -111,7 +65,7 @@ export class DevicesService {
         const json = await response.json();
         return {
           result: 'success',
-          device: normalizeDeviceModel(json.device),
+          device: json.device,
         };
       } else {
         return handleError(await response.text());
